refactor: 포인트 도메인 폴더 분리
