+++ src/main/java/kr/tennispark/members/user/infrastructure/repository/PointHistoryRepository.java
@@ -1,6 +1,6 @@
 package kr.tennispark.members.user.infrastructure.repository;
 
-import kr.tennispark.members.common.domain.entity.association.PointHistory;
+import kr.tennispark.point.common.domain.entity.PointHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PointHistoryRepository extends JpaRepository<PointHistory, Long> {
